+++ UCSO.h
@@ -4,7 +4,7 @@
 
 class UCSO : public CargoVessel {
 public:
-	UCSO(OBJHANDLE hObj, int fmodel) : CargoVessel(hObj, fmodel) {};
+	UCSO(OBJHANDLE hObj, int fmodel) : CargoVessel(hObj, fmodel) { LoadConfig(); };
 	~UCSO() {}
 	void clbkSetClassCaps(FILEHANDLE cfg);
 	void clbkLoadStateEx(FILEHANDLE scn, void* status);
@@ -13,12 +13,6 @@ public:
 	void clbkSaveState(FILEHANDLE scn);
 
 private:
-	enum UnpackMode {
-		LANDED = 0,
-		DELAYED,
-		MANUAL
-	};
-
 	double timer = 0;
 	bool landing = false;
 	bool timing = false;